main.py
#!/usr/bin/env python3
"""
Flickd Smart Tagging Engine
A Python FastAPI application for AI-powered image tagging
"""
from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.staticfiles import StaticFiles
from fastapi.responses import HTMLResponse, JSONResponse
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
import logging
import os
import io
from PIL import Image
import base64
import json
from typing import List, Dict, Any
import asyncio
from datetime import datetime

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('flickd_tagging.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

app = FastAPI(
    title="Flickd Smart Tagging Engine",
    description="AI-powered image tagging system with confidence scoring",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Mount static files
app.mount("/static", StaticFiles(directory="static"), name="static")

class ImageTagger:
    def __init__(self):
        self.model_loaded = False
        logger.info("Initializing ImageTagger")
    
    async def analyze_image(self, image: Image.Image) -> Dict[str, Any]:
        """Analyze image and return tags with confidence scores"""
        try:
            logger.info("Starting image analysis")
            
            # Simulate AI analysis with realistic tags and confidence scores
            # In a real implementation, you would use actual AI models here
            await asyncio.sleep(1)  # Simulate processing time
            
            # Basic image analysis
            width, height = image.size
            aspect_ratio = width / height
            
            # Generate tags based on image properties and simulated AI analysis
            tags = []
            
            # Color analysis simulation
            dominant_colors = self._analyze_colors(image)
            for color, confidence in dominant_colors:
                tags.append({
                    "tag": f"{color}_dominant",
                    "confidence": confidence,
                    "category": "color"
                })
            
            # Object detection simulation
            objects = self._simulate_object_detection(image)
            for obj, confidence in objects:
                tags.append({
                    "tag": obj,
                    "confidence": confidence,
                    "category": "object"
                })
            
            # Scene analysis simulation
            scenes = self._simulate_scene_analysis(aspect_ratio, width, height)
            for scene, confidence in scenes:
                tags.append({
                    "tag": scene,
                    "confidence": confidence,
                    "category": "scene"
                })
            
            # Sort by confidence
            tags.sort(key=lambda x: x["confidence"], reverse=True)
            
            logger.info(f"Generated {len(tags)} tags for image")
            return {
                "tags": tags,
                "image_info": {
                    "width": width,
                    "height": height,
                    "aspect_ratio": round(aspect_ratio, 2),
                    "format": image.format or "Unknown"
                },
                "processing_time": 1.0
            }
            
        except Exception as e:
            logger.error(f"Error analyzing image: {str(e)}")
            raise HTTPException(status_code=500, detail=f"Image analysis failed: {str(e)}")
    
    def _analyze_colors(self, image: Image.Image) -> List[tuple]:
        """Simulate color analysis"""
        colors = [
            ("blue", 0.85),
            ("green", 0.72),
            ("red", 0.68),
            ("yellow", 0.45),
            ("purple", 0.32)
        ]
        return colors[:2]  # Return top 2 colors
    
    def _simulate_object_detection(self, image: Image.Image) -> List[tuple]:
        """Simulate object detection"""
        objects = [
            ("person", 0.92),
            ("car", 0.78),
            ("building", 0.85),
            ("tree", 0.73),
            ("sky", 0.89),
            ("road", 0.67),
            ("animal", 0.54),
            ("food", 0.61)
        ]
        # Return 3-5 random objects
        import random
        return random.sample(objects, random.randint(3, 5))
    
    def _simulate_scene_analysis(self, aspect_ratio: float, width: int, height: int) -> List[tuple]:
        """Simulate scene analysis"""
        scenes = []
        
        if aspect_ratio > 1.5:
            scenes.append(("landscape", 0.88))
            scenes.append(("panoramic", 0.76))
        elif aspect_ratio < 0.8:
            scenes.append(("portrait", 0.91))
            scenes.append(("vertical", 0.82))
        else:
            scenes.append(("square", 0.79))
        
        if width > 1920:
            scenes.append(("high_resolution", 0.95))
        
        scenes.append(("outdoor", 0.71))
        scenes.append(("daytime", 0.83))
        
        return scenes

# Initialize the tagger
tagger = ImageTagger()

@app.get("/", response_class=HTMLResponse)
async def read_root():
    """Serve the main web interface"""
    try:
        with open("static/index.html", "r") as f:
            return HTMLResponse(content=f.read())
    except FileNotFoundError:
        return HTMLResponse(content="""
        <html>
            <head><title>Flickd Smart Tagging Engine</title></head>
            <body>
                <h1>Flickd Smart Tagging Engine</h1>
                <p>Web interface not found. Please ensure static/index.html exists.</p>
                <p>API is available at <a href="/docs">/docs</a></p>
            </body>
        </html>
        """)

@app.post("/api/analyze")
async def analyze_image(file: UploadFile = File(...)):
    """Analyze uploaded image and return tags with confidence scores"""
    try:
        logger.info(f"Received image upload: {file.filename}")
        
        # Validate file type
        if not file.content_type.startswith('image/'):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read and process image
        contents = await file.read()
        image = Image.open(io.BytesIO(contents))
        
        # Convert to RGB if necessary
        if image.mode != 'RGB':
            image = image.convert('RGB')
        
        # Analyze image
        result = await tagger.analyze_image(image)
        
        # Add metadata
        result["filename"] = file.filename
        result["timestamp"] = datetime.now().isoformat()
        
        logger.info(f"Successfully analyzed image: {file.filename}")
        return JSONResponse(content=result)
        
    except Exception as e:
        logger.error(f"Error processing upload: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "healthy", "timestamp": datetime.now().isoformat()}

@app.get("/api/stats")
async def get_stats():
    """Get system statistics"""
    return {
        "model_status": "loaded" if tagger.model_loaded else "not_loaded",
        "supported_formats": ["JPEG", "PNG", "GIF", "BMP", "TIFF"],
        "max_file_size": "10MB",
        "version": "1.0.0"
    }

def main():
    """Main entry point for the application"""
    logger.info("Starting Flickd Smart Tagging Engine")
    uvicorn.run(app, host="0.0.0.0", port=int(os.getenv("PORT", 8000)))

if __name__ == "__main__":
    main()
